+++ CHANGES.rst
@@ -4,9 +4,7 @@ Changelog
 5.2.3 (unreleased)
 ------------------
 
-- Fix race condition affecting ``DB.open`` by
-  moving the `setLastTID` call after invalidation processing as
-  required by ZODB>=5.6.0
+- Fix race condition affecting ``DB.open``
   `#166 <https://github.com/zopefoundation/ZEO/issues/166>`_.